+++ ov_core/src/feat/Feature.h
@@ -53,19 +53,19 @@ namespace ov_core {
         std::unordered_map<size_t, std::vector<Eigen::VectorXf>> uvs_norm;
 
         /// Timestamps of each UV measurement (mapped by camera ID)
-        std::unordered_map<size_t, std::vector<double>> timestamps;
+        std::unordered_map<size_t, std::vector<float>> timestamps;
 
         /// What camera ID our pose is anchored in!! By default the first measurement is the anchor.
         int anchor_cam_id = -1;
 
         /// Timestamp of anchor clone
-        double anchor_clone_timestamp;
+        float anchor_clone_timestamp;
 
         /// Triangulated position of this feature, in the anchor frame
-        Eigen::Vector3d p_FinA;
+        Eigen::Vector3f p_FinA;
 
         /// Triangulated position of this feature, in the global frame
-        Eigen::Vector3d p_FinG;
+        Eigen::Vector3f p_FinG;
 
 
         /**
@@ -76,7 +76,7 @@ namespace ov_core {
          *
          * @param valid_times Vector of timestamps that our measurements must occur at
          */
-        void clean_old_measurements(std::vector<double> valid_times);
+        void clean_old_measurements(std::vector<float> valid_times);
 
         /**
          * @brief Remove measurements that are older then the specified timestamp.
@@ -85,7 +85,7 @@ namespace ov_core {
          *
          * @param timestamp Timestamps that our measurements must occur after
          */
-        void clean_older_measurements(double timestamp);
+        void clean_older_measurements(float timestamp);
 
     };
